Add files via upload
The goal of this project is to use the MNIST dataset to train a neural network model for the classification of handwritten numbers. The objective is to develop a system that, given input photos, can reliably identify numbers 0 through 9. One well-known machine learning benchmark, the MNIST dataset, offers a collection of 28x28 grayscale pictures of handwritten numbers. Preprocessing the data, training the model, assessing it, and implementing a prediction script are all part of the project. It provides an overview of key ideas in neural networks, such as activation functions, thick layers, and metrics for measuring model evaluation.